@@ -1,115 +1,80 @@
 {
   "name": "tout-ce-qui-bouge",
-  "engines": {
-    "node": "18.15.0"
-  },
+  "type": "module",
   "scripts": {
-    "dev:webpack": "webpack --watch  --mode development",
-    "dev:eleventy": "  npx @11ty/eleventy  --incremental --serve --quiet",
-    "debug:eleventy": "npx @11ty/eleventy --incremental",
+    "dev:webpack": "npx webpack --watch  --mode development",
+    "dev:copy": "npx cpy-cli  --flat \"./src/posts/**/*.{png,webp,gif,mp4,jpg,jpeg}\" \"./src/assets/images/*\" \"./src/assets/UI/*\"  ./src/assets/imagesToProcess/",
+    "dev:eleventy": "npx @11ty/eleventy --incremental  --serve --quiet",
     "dev:postcss": "postcss src/assets/css/main.css -o dist/assets/css/main.css --watch",
     "dev": "npm-run-all --parallel dev:* --print-label",
-    "threader": "  set DEBUG=threader:* & npx tsx watch ./src/heroPages/threads/threads_test.ts",
-    "build:webpack": "webpack  --mode production",
+    "build:copy": "npx cpy-cli  --flat \"./src/posts/**/*.{png,webp,gif,mp4,jpg,jpeg}\" \"./src/assets/images/*\" \"./src/assets/UI/*\"  ./src/assets/imagesToProcess/",
+    "build:webpack": "npx webpack  --mode production",
+    "build:eleventy": " npx @11ty/eleventy --incremental --quiet",
     "build:postcss": "cross-env NODE_ENV=production postcss --verbose src/assets/css/main.css -o dist/assets/css/main.css",
-    "build:eleventy": "cross-env NODE_ENV=production npx @11ty/eleventy  --quiet --incremental",
+    "build:search": "npx pagefind --site dist ",
     "build": "run-s build:* --print-label"
   },
   "devDependencies": {
-    "@11ty/eleventy": "2.0.1",
+    "@11ty/eleventy": "github:11ty/eleventy",
     "@11ty/eleventy-fetch": "4.0.0",
-    "@11ty/eleventy-img": "4.0.1 ",
+    "@11ty/eleventy-img": "^4.0.2",
     "@11ty/eleventy-navigation": "^0.3.5",
     "@11ty/eleventy-plugin-directory-output": "^1.0.1",
     "@11ty/eleventy-plugin-rss": "github:baptiste-roullin/eleventy-plugin-rss",
-    "@babel/core": "^7.20.2",
-    "@babel/preset-env": "^7.20.2",
-    "@sindresorhus/slugify": "1.1.0",
-    "@tailwindcss/forms": "^0.5.3",
-    "@tailwindcss/typography": "^0.5.8",
-    "@types/elasticlunr": "^0.9.5",
+    "@tailwindcss/forms": "*",
+    "@tailwindcss/typography": "*",
     "@types/luxon": "^3.1.0",
-    "@types/markdown-it": "^12.2.3",
+    "@types/markdown-it": "^13.0.5",
     "@types/markdown-it-container": "^2.0.5",
-    "@types/node": "^18.11.9",
+    "@types/node": "^20.8.10",
     "@types/nunjucks": "^3.2.1",
-    "@types/remove-markdown": "^0.3.1",
     "@types/webpack": "^5.28.0",
     "autoprefixer": "^10.4.13",
-    "babel-loader": "^9.1.0",
-    "child_process": "^1.0.2",
-    "copy-webpack-plugin": "^11.0.0",
+    "canvas": "2.11.2",
+    "cpy-cli": "^5.0.0",
     "cross-env": "^7.0.3",
-    "css-loader": "^6.7.1",
     "cssnano": "^6.0.0  ",
     "d3-color": "^3.1.0",
     "debug": "^4.3.4",
     "dotenv": "^16.0.3",
     "eleventy-plugin-embed-everything": "^1.14",
-    "ffmpeg-static": "^5.1.0",
+    "esm-seedrandom": "^3.0.5",
     "full-icu": "^1.5.0",
     "js-yaml": "^4.1.0",
-    "linkedom": "^0.14.19",
+    "linkedom": "^0.16.1",
     "lodash.clonedeep": "^4.5.0",
     "luxon": "^3.1.0",
-    "markdown-it": "^13.0.1",
+    "markdown-it": "^14.0.0",
     "markdown-it-anchor": "^8.6.5",
     "markdown-it-attrs": "^4.1.4",
     "markdown-it-blockquote-cite": "github:baptiste-roullin/markdown-it-blockquote-cite",
     "markdown-it-bracketed-spans": "^1.0.1",
-    "markdown-it-container": "^3.0.0",
-    "markdown-it-footnote": "^3.0.3",
+    "markdown-it-container": "^4.0.0",
+    "markdown-it-footnote": "^4.0.0",
     "markdown-it-highlightjs": "^4.0.1",
     "markdown-it-image-figures": "^2.1.0",
-    "metascraper": "^5.31.3",
-    "metascraper-author": "^5.32.1",
-    "metascraper-clearbit": "^5.32.1",
-    "metascraper-date": "^5.32.1",
-    "metascraper-description": "^5.32.1",
-    "metascraper-image": "^5.32.1",
-    "metascraper-logo": "^5.32.1",
-    "metascraper-publisher": "^5.32.1",
-    "metascraper-title": "^5.32.1",
-    "metascraper-url": "^5.32.1",
+    "node-polyfill-webpack-plugin": "^3.0.0",
     "npm-run-all": "^4.1.5",
-    "p-map": "^5.5.0",
+    "p-map": "^7.0.0",
+    "pagefind": "^1.0.4",
     "postcss": "^8.4.19",
-    "postcss-cli": "^10.0.0",
-    "postcss-hash": "^3.0.0",
-    "postcss-import": "^15.0.0",
-    "random-seed": "^0.3.0",
-    "remove-markdown": "^0.5.0",
-    "simple-nunjucks-loader": "^3.2.0",
-    "style-loader": "^3.3.1",
-    "tailwindcss": "^3.2.4",
-    "tailwindcss-textshadow": "^2.1.3",
-    "ts-loader": "^9.4.1",
-    "ts-node": "^10.9.1",
-    "twitter-api-sdk": "^1.2.1",
-    "webpack": "^5.75.0",
+    "postcss-cli": "^11.0.0",
+    "postcss-hash": "github:keithamus/postcss-hash",
+    "postcss-import": "^16.0.0",
+    "simple-nunjucks-loader": "github:baptiste-roullin/nunjucks-loader",
+    "tailwindcss": "*",
+    "webpack": "^5.89.0",
     "webpack-assets-manifest": "^5.1.0",
-    "webpack-cli": "^5.0.1",
-    "zotero-api-client": "^0.40.0"
-  },
-  "optionalDependencies": {
-    "canvas": "^2.10.2"
+    "webpack-cli": "^5.1.4",
+    "zotero-api-client": "^0.43.1"
   },
   "dependencies": {
-    "elasticlunr": "github:weixsong/elasticlunr.js#v0.9.6",
-    "photoswipe": "^5.3.4",
-    "text-metrics": "^3.0.0"
+    "photoswipe": "^5.4.3",
+    "text-metrics": "^4.0.1"
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
     "not dead"
-  ],
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": "> 0.25%, not dead"
-      }
-    ]
   ]
 }
