@@ -1,14 +1,20 @@
-const meta = require('./src/_data/meta')
+//TODO : refaire marcher
+//import * as cssnano from 'cssnano'
 
-module.exports = {
-  plugins: [
-    require('postcss-import'),
-    require('tailwindcss/nesting'),
-    require('tailwindcss')({ config: './tailwind.config.js' }),
-    require('autoprefixer'),
-    ...(meta.env === "production" ? [require('cssnano')] : []),
-    require('postcss-hash')({
-      manifest: './src/_data/hashes_css.json',
-    }),
-  ]
+//import meta from './src/_data/meta.js'
+
+//const nano = (meta.env === "production" ?  : ""),
+
+
+
+export default {
+  plugins: {
+    'postcss-import': {},
+    'tailwindcss/nesting': {},
+    tailwindcss: { config: './tailwind.config.js' },
+    "autoprefixer": {},
+    //TODO : refaire marcher
+    //    "postcss-hash": {},
+    "cssnano": {}
+  }
 }
