@@ -1,2 +1,2 @@
-import { md } from '../config/markdown'
+import md from '../markdown.js'
 export default function (markdownString) { return md.renderInline(markdownString) }
