@@ -1,8 +1,14 @@
 
-import { truchetItem, truchetList } from '../features/truchet/truchet_shortcode'
-import { slugifyFilter as slugify } from '../filters/slugify'
+import dotenv from 'dotenv'
+dotenv.config()
 
-export const shortcodes = {
+import { truchetItem, truchetList } from '../features/truchet/truchet_shortcode.js'
+import slugify from '../filters/slugify.js'
+import zotero from '../features/zotero/zotero.js'
+import meta from '../_data/meta.js'
+
+export default {
+  zotero: (!meta.zotero ? async () => "[ZOTÉRO DÉSACTIVÉ]" : zotero),
   truchetItem: truchetItem,
   truchetList: truchetList,
   heading: function (level, className, label) {
@@ -30,6 +36,7 @@ export const shortcodes = {
               </a>
     </div>  `
   },
+
   /**
    * ===== SVGs =====
  * This shortcode is used in layouts and can be used in .md content.
